+++ Core/RuntimeBlock.hpp
@@ -1,9 +1,9 @@
-#include "Compiler.h"
+#include "Compiler.hpp"
 #include "Error.hpp"
-#include "GarbageCollector.h"
+#include "GarbageCollector.hpp"
 #include "List.hpp"
 #include "PortableMemPool.hpp"
-#include "Types.h"
+#include "Types.hpp"
 #include <CL/sycl.hpp>
 #include <atomic>
