+++ Core/SExpr.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Types.h"
+#include "Types.hpp"
 #include <memory>
 #include <string>
 #include <vector>
